pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "pdf-ai-context-extractor"
version = "0.7.0"
description = "Extract structured data from PDF tables"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "tabula-py>=2.0.0",
    "pandas>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
